Update Read me.txt
The simple calculator apk code is developed and uploaded 
In this reposotary. 
This code uses a flutters dart language. 
To use of this code please download and place the code
In the lib folder of the flutter files. 
The apk link is also provided in this reposotary 
you can download and install the app in your mobile 
Like installing apps from play store.